pyproject.toml
[tool.poetry]
name = "webcamrecorder"
version = "0.1.0"
license = "MIT"
description = ""
authors = ["kwhr kzk <kwhrkzk83.gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-contrib-python = "^4.7.0.72"
python-dotenv = "^1.0.0"
injector = "^0.20.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"